getAccessToken refactored into ContextProvider #79
