+++ frontend/src/App.tsx
@@ -4,16 +4,19 @@ import ImageUpload from "./Components/ImageUpload/imageUpload";
 import ImageList from "./Components/ImageList/imageList";
 import { Route, Switch } from "react-router-dom";
 import UserProfile from "./Components/UserProfile/userProfile";
+import AuthenticationProvider from "./Components/AuthenticationProvider/authenticationProvider";
 
 function App() {
   return (
     <div className="App">
       <header className="App-header">
-        <Switch>
-          <Route path="/" exact component={ImageList} />
-          <Route path="/create" exact component={ImageUpload} />
-          <Route path="/user/:userId" component={UserProfile} />
-        </Switch>
+        <AuthenticationProvider>
+          <Switch>
+            <Route path="/" exact component={ImageList} />
+            <Route path="/create" exact component={ImageUpload} />
+            <Route path="/user/:userId" component={UserProfile} />
+          </Switch>
+        </AuthenticationProvider>
       </header>
     </div>
   );
