+++ frontend/src/Components/ImageUpload/imageUpload.tsx
@@ -1,23 +1,18 @@
 import { Button, TextField, Paper, LinearProgress } from "@material-ui/core";
-import { useEffect, useState } from "react";
+import { useContext, useEffect, useState } from "react";
 import "./imageUpload.css";
 import { useHistory } from "react-router-dom";
-import { useAccount, useIsAuthenticated, useMsal } from "@azure/msal-react";
-import { loginRequest } from "../../Constants/authConfig";
-import {
-  InteractionRequiredAuthError,
-  SilentRequest,
-} from "@azure/msal-browser";
+import { useIsAuthenticated } from "@azure/msal-react";
 import { backendURL } from "../../Constants/backendConfig";
+import { AuthenticationContext, AuthenticationContextType } from "../AuthenticationProvider/authenticationProvider";
 
 export default function ImageUpload() {
   const [file, setFile] = useState<File>();
   const [description, setDescription] = useState('');
   const [loading, setLoading] = useState(false);
-  const { instance, accounts } = useMsal();
-  const account = useAccount(accounts[0] || {})!;
   const history = useHistory();
   const isAuthenticated = useIsAuthenticated();
+  const { getAccessToken } = useContext(AuthenticationContext) as AuthenticationContextType
 
 
   useEffect(() => {
@@ -28,27 +23,6 @@ export default function ImageUpload() {
   function checkAuth() {
     if (isAuthenticated === false) {
       history.push('/')
-    }
-  }
-
-  async function getAccessToken() {
-    const silentRequest: SilentRequest = {
-      account: account,
-      ...loginRequest,
-    };
-    try {
-      const resp = await instance
-        .acquireTokenSilent(silentRequest);
-      if (resp.accessToken) {
-        return resp.accessToken;
-      }
-    } catch (error) {
-      if (error instanceof InteractionRequiredAuthError) {
-        // fallback to interaction when silent call fails
-        instance.acquireTokenPopup(silentRequest).then((response) => {
-          return response.accessToken
-        });
-      }
     }
   }
 
@@ -71,11 +45,9 @@ export default function ImageUpload() {
       },
       body: formData,
     })
-      .then((response) => {
+      .then(() => {
         history.push('/')
       });
-    // setTimeout(() => {
-    // }, 1000);
   }
 
   return (
@@ -84,8 +56,8 @@ export default function ImageUpload() {
         {file != null ? (
           <img id="img" src={URL.createObjectURL(file)} alt="" />
         ) : (
-            ""
-          )}
+          ""
+        )}
       </div>
       <Button variant="contained" component="label">
         Select Image
