+++ frontend/src/Constants/backendConfig.ts
@@ -1,3 +1,3 @@
 // URL of the backend server
-export const backendURL: string = "https://localhost:5001"
-  
+export const backendURL: string = "https://soen341.grypr.cf:5001"
+// export const backendURL: string = "https://localhost:5001"
